catalog/callbacks/finops_before_model/requirements.txt
google-adk>=1.9.0
